stateful_agent/pyproject.toml
[project]
name = "stateful-agent"
version = "0.1.0"
description = "A sophisticated agent system with persistent state management using SQLite and ChromaDB"
authors = [
    {name = "Sangdae Nam", email = "nsd96@berkeley.edu"}
]
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["agent", "langchain", "chromadb", "sqlite", "vector-database"]
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.14",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "arxiv>=2.1.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4>=4.13.0",
    "black>=23.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb>=0.6.3",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "contourpy==1.3.1",
    "cryptography==44.0.2",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "dill==0.3.9",
    "distro==1.9.0",
    "durationpy==0.9",
    "dynaconf==3.2.10",
    "fastapi==0.115.11",
    "feedparser==6.0.11",
    "filelock==3.17.0",
    "flatbuffers==25.2.10",
    "flask>=3.0.0",
    "fonttools==4.56.0",
    "frozenlist==1.5.0",
    "fsspec==2025.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.38.0",
    "googleapis-common-protos==1.69.1",
    "greenlet==3.1.1",
    "grpcio==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.27.0",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.29.2",
    "humanfriendly==10.0",
    "hyperdock-fileio>=0.5.3",
    "hyperpocket>=0.5.2",
    "hyperpocket-langchain>=0.5.2",
    "idna==3.10",
    "importlib-metadata==8.5.0",
    "importlib-resources==6.5.2",
    "iniconfig==2.1.0",
    "isort>=5.0",
    "jinja2==3.1.6",
    "jiter==0.8.2",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.8",
    "kubernetes==32.0.1",
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.20",
    "langchain-core==0.3.49",
    "langchain-openai>=0.3.1",
    "langchain-text-splitters==0.3.7",
    "langsmith==0.3.13",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.2.0",
    "multiprocess==0.70.17",
    "mypy>=1.0",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy>=1.26.0",
    "oauthlib==3.2.2",
    "onnxruntime==1.15.1",
    "openai==1.70.0",
    "opentelemetry-api==1.30.0",
    "opentelemetry-exporter-otlp-proto-common==1.30.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.30.0",
    "opentelemetry-instrumentation==0.51b0",
    "opentelemetry-instrumentation-asgi==0.51b0",
    "opentelemetry-instrumentation-fastapi==0.51b0",
    "opentelemetry-proto==1.30.0",
    "opentelemetry-sdk==1.30.0",
    "opentelemetry-semantic-conventions==0.51b0",
    "opentelemetry-util-http==0.51b0",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pathspec==0.12.1",
    "pillow==11.1.0",
    "platformdirs==4.3.7",
    "pluggy==1.5.0",
    "posthog==3.19.0",
    "propcache==0.3.1",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.8.1",
    "pygithub==2.6.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pynacl==1.5.0",
    "pyparsing==3.2.1",
    "pypdf>=5.4.0",
    "pypdf2==3.0.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pytest>=7.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.0",
    "python-multipart==0.0.20",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "ratelimit==2.2.1",
    "redis==5.2.1",
    "regex==2024.11.6",
    "requests>=2.30.0",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rsa==4.9",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==3.4.1",
    "setuptools==75.8.0",
    "sgmllib3k==1.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.38",
    "starlette==0.46.1",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "threadpoolctl==3.5.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.0",
    "toml==0.10.2",
    "torch==2.2.2",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2025.1",
    "tzlocal>=5.0.0",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
    "wrapt==1.17.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.hatch.build.targets.wheel]
packages = ["stateful_agent"]